--- tencentcloud/data_source_tc_sqlserver_datasource_ins_attribute_test.go
@@ -1,33 +1,0 @@
-package tencentcloud
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-)
-
-// go test -i; go test -test.run TestAccTencentCloudSqlserverDatasourceInsAttributeDataSource_basic -v
-func TestAccTencentCloudSqlserverDatasourceInsAttributeDataSource_basic(t *testing.T) {
-	t.Parallel()
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() {
-			testAccPreCheck(t)
-		},
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccSqlserverDatasourceInsAttributeDataSource,
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckTencentCloudDataSourceID("data.tencentcloud_sqlserver_datasource_ins_attribute.datasource_ins_attribute"),
-					resource.TestCheckResourceAttrSet("data.tencentcloud_sqlserver_datasource_ins_attribute.datasource_ins_attribute", "instance_id"),
-				),
-			},
-		},
-	})
-}
-
-const testAccSqlserverDatasourceInsAttributeDataSource = `
-data "tencentcloud_sqlserver_datasource_ins_attribute" "datasource_ins_attribute" {
-  instance_id = "mssql-gyg9xycl"
-}
-`
