+++ website/tencentcloud.erb
@@ -1994,13 +1994,13 @@
                                     <a href="/docs/providers/tencentcloud/d/sqlserver_cross_region_zone.html">tencentcloud_sqlserver_cross_region_zone</a>
                                 </li>
                                 <li>
-                                    <a href="/docs/providers/tencentcloud/d/sqlserver_datasource_ins_attribute.html">tencentcloud_sqlserver_datasource_ins_attribute</a>
-                                </li>
-                                <li>
                                     <a href="/docs/providers/tencentcloud/d/sqlserver_db_charsets.html">tencentcloud_sqlserver_db_charsets</a>
                                 </li>
                                 <li>
                                     <a href="/docs/providers/tencentcloud/d/sqlserver_dbs.html">tencentcloud_sqlserver_dbs</a>
+                                </li>
+                                <li>
+                                    <a href="/docs/providers/tencentcloud/d/sqlserver_ins_attribute.html">tencentcloud_sqlserver_ins_attribute</a>
                                 </li>
                                 <li>
                                     <a href="/docs/providers/tencentcloud/d/sqlserver_instance_param_records.html">tencentcloud_sqlserver_instance_param_records</a>
